\name{wBoot-package}
\alias{wBoot-package}
\alias{wBoot}
\docType{package}

\title{
wBootstrap Routines
}

\description{
Supplies bootstrap alternatives to traditional hypothesis-test and confidence-interval procedures such as one-sample and two-sample inferences for means, standard deviations, and proportions; simple linear regression; and more. Suitable for general audiences, including individual and group users, introductory statistics courses, and more advanced statistics courses that desire an introduction to bootstrap methods.
}

\details{
\tabular{ll}{
Package: \tab wBoot\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-09-21\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Neil A. Weiss

Maintainer: Neil A. Weiss <neil.a.weiss@asu.edu> 

}

\keyword{ package, datasets, graphs }

\concept{ bootstrap methods }
\concept{ bootstrap }
\concept{ simulation }
\concept{ inference }